the app runs
