+++ ios/Shared/OurHomeApp.swift
@@ -17,9 +17,11 @@ extension EnvironmentValues {
 
 @main
 struct OurHomeApp: App {
+#if !os(watchOS)
   @UIApplicationDelegateAdaptor
   private var appDelegate: AppDelegate
-  
+#endif
+
   @StateObject
   var appState = AppState()
     
@@ -43,10 +45,12 @@ struct OurHomeApp: App {
             }
         }
         .onAppear() {
+#if !os(watchOS)
           self.appDelegate.pushNotificationSync = appState.pushNotificationSync
+#endif
           appState.loadUser()
           
-          UIApplication.shared.registerForRemoteNotifications()
+          NotificationProvider.registerForRemoteNotifications()
           
           if ProcessInfo.processInfo.environment["FAKE_PUSH"] == "1" {
             appState.notificationProvider.showBuzzerNotification(delayed: true)
@@ -96,14 +100,5 @@ extension OurHomeApp {
     case .unlatchDoor:
       trigger(action: .unlatchDoor)
     }
-  }
-}
-
-class AppDelegate: NSObject, UIApplicationDelegate, ObservableObject {
-  weak var pushNotificationSync: PushNotificationSync?
-  
-  func application(_ application: UIApplication, didRegisterForRemoteNotificationsWithDeviceToken deviceToken: Data) {
-    let tokenParts = deviceToken.map { data in String(format: "%02.2hhx", data) }
-    self.pushNotificationSync?.deviceToken = tokenParts.joined()
   }
 }
